added login

+++ misc/import.py
@@ -66,6 +66,26 @@ qas = [{
         },  
         {
         "category": "sonst",
+        "studiengang": [],
+        "q_de": "Wie funktioniert ein Auslandssemester?", 
+        "q_en": "",
+        "a_de": '''[Studienberatung]({{url_for('studienberatungbase')}}) International Office der Universität Freiburg: \n\n Fahnenbergplatz, 79085 Freiburg \n Kurzinformation (pdf)  \n [Webseite](https://www.international.uni-freiburg.de/) \n Prof. Dr. Patrick Dondl \n _Auslandsbeautragter des Mathematischen Instituts_\n [Internetseite des Auslandsbeauftragten](https://home.mathematik.uni-freiburg.de/erasmus) Hermann-Herder-Straße 10, Raum 217 Tel. +49 761 203-5642 E-Mail: erasmus@math.uni-freiburg.de Sprechstunde: Mo 12:15–13:45 Uhr''',         
+        "a_en": "",
+        "rang": 11,
+        "kommentar": ""
+        },  
+        {
+        "category": "sonst",
+        "studiengang": [],
+        "q_de": "Wie melde ich mich zu einer Prüfung an?", 
+        "q_en": "",
+        "a_de": "Kurzanleitung: \n1. In [HISinOne](https://campus.uni-freiburg.de/) _'Mein Studium > Studienplaner mit Modulplan'_ (ggf. mit richtigem Studiengang) auswählen.\n2. Links oben das richtige Semester auswählen und rechts oben bei Veranstaltungen _'keine'_ und bei Prüfungs-/Studienleistungen _'alle'_ anklicken. \n3. Dann in der Liste das gewünschte Modul bzw. die gewünschte Veranstaltung suchen. \n4. Auf _'Prüfung anmelden'_ bzw. _'Studienleistung registrieren'_ gehen. \n5. Gegebenenfalls noch die richtige Gruppe auswählen. ",         
+        "a_en": "",
+        "rang": 12,
+        "kommentar": ""
+        },  
+        {
+        "category": "sonst",
         "studiengang": ["med", "mederw"],
         "q_de": "Wer ist das LLPA?", 
         "q_en": "",
@@ -75,11 +95,16 @@ qas = [{
         "kommentar": ""
         }  ]
 
-category.delete_many({})
-qa.delete_many({})
-    
-for x in qas:
-    qa.insert_one(x)
-for x in categories:
-    category.insert_one(x)
+
+print("Dieser Vorgang löscht alle Einträge in category und qa!")
+print("Wirklch fortfahren? [y/N]")
+input = input()
+if input == "y":
+  category.delete_many({})
+  qa.delete_many({})
+      
+  for x in qas:
+      qa.insert_one(x)
+  for x in categories:
+      category.insert_one(x)
